Split store into slices -- WORKING
This splits the existing store into dedicated slices. This is the 1st step towards refactoring the store to be free of side effects.

In the next step, functions will be extracted from the store into either Components, Hooks, or Utility functions.
