+++ src/components/ProfileManager.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import useStore from '../store';
+import useStore from '../store/index';
 import type { Profile } from '../types/Profile';
 import { toast } from 'sonner';
