+++ src/lib/debugLog.ts
@@ -1,4 +1,4 @@
-import useStore from '../store';
+import useStore from '../store/index';
 
 export function debugLog(...args: any[]) {
   const debugMode = useStore.getState().debugMode;
